src/server/requirements-worker.txt
hidapi==0.7.99.post21
modbus-tk==0.5.4
pyserial==3.1
pyzmq==16.0.2
schematics==1.1.1
six==1.10.0
watchdog==0.8.3
rpi.gpio==0.6.3
